use Uint32List? instead

+++ lib/src/platform_file.dart
@@ -52,7 +52,8 @@ class PlatformFile {
   /// Byte data for this file. Particularly useful if you want to manipulate its data
   /// or easily upload to somewhere else.
   /// [Check here in the FAQ](https://github.com/miguelpruivo/flutter_file_picker/wiki/FAQ) an example on how to use it to upload on web.
-  final Uint8List? bytes;
+  //final Uint8List? bytes;
+  final Uint32List? bytes;
 
   /// File content as stream
   final Stream<List<int>>? readStream;
